Move Argo file changes stuff into Calamari contracts (#1914)
