+++ source/Calamari.Contracts/ArgoCD/FileHash.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace Octopus.Calamari.Contracts.ArgoCD;
+
+public record FileHash(string FilePath, string Hash);
