+++ source/Calamari.Contracts/ArgoCD/FileJsonPatch.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace Octopus.Calamari.Contracts.ArgoCD;
+
+public record FileJsonPatch(string FilePath, string JsonPatch);
