+++ source/Calamari.Contracts/ArgoCD/SourceFileChanges.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Octopus.Calamari.Contracts.ArgoCD;
+
+public record SourceFileChanges(
+    string? CommitSha,
+    DateTimeOffset? CommitTimestamp,
+    int SourceIndex,
+    List<FileHash> ReplacedFiles,
+    List<FileJsonPatch> PatchedFiles);
