+++ source/Calamari/ArgoCD/Conventions/ManifestTemplating/ManifestUpdateResult.cs
@@ -1,6 +1,7 @@
 #nullable enable
 using System;
 using System.Collections.Generic;
+using Octopus.Calamari.Contracts.ArgoCD;
 
 namespace Calamari.ArgoCD.Conventions.ManifestTemplating;
